requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-telegram-bot==20.8
python-dotenv==1.0.1
pydantic==2.6.4
PyYAML==6.0.2
python-multipart==0.0.20
